Functional enrichment analysis

Performs gene ontology and impacted pathway enrichment analysis with a list of gene names and their fold-change.

find_impacted_pathways(
  gene_file = NULL,
  reference_file = NULL,
  enrichment_tool = c("WebGestaltR", "ROntoTools", "enrichR"),
  enrichment_database = c("KEGG", "Panther", "Reactome"),
  is_output = FALSE,
  output_dir = ".",
  ...
)

Arguments

gene_file
A data frame or the path of a .tsv file containing a list of genes, their fold-change, p-value and adjusted p-value. Column names should be gene, logFC, pval and padj respectively.

reference_file
A data frame containing all the genes that were used as input for differential expression. Column name should be gene. If not provided the human protein-coding genome will be used as background genes.

enrichment_tool
Enrichment tool to use. WebGestaltR, ROntoTools and enrichR is implemented. Use one or more of the tools.

enrichment_database
Name of the database for enrichment. User can specify one or more database names. Check scFlow::list_databases() for available database alias.

is_output
If TRUE a folder will be created and results of enrichment analysis will be saved otherwise a R list will be returned. Default FALSE.

output_dir
Path for the output directory. Default is current directory.

Value

enrichment_result a list of list containing enrichment outputs from different enrichment tools.

See also

Examples

enrichment_result <- find_impacted_pathways(
  gene_file = paste(system.file("extdata", package = "scFlowData"), "/",
    "de_result_table.tsv",
    sep = ""
  ),
  enrichment_method = "ORA"
)

#> Error in `$<-.data.frame`(`*tmp*`, "db_alias", value = c("KEGG", "NCI", "Panther", "Reactome")): replacement has 4 rows, data has 0