Functional enrichment analysis using enrichR

Performs impacted pathway analysis with a list of genes.

pathway_analysis_enrichr(
  gene_file = NULL,
  enrichment_database = c("GO_Molecular_Function_2021", "GO_Cellular_Component_2021",
    "GO_Biological_Process_2021", "KEGG_2021_Human", "WikiPathways_2021_Human",
    "Reactome_2022")
)

Arguments

gene_file

A data frame containing a list of significant genes with column name gene or a vector of significant genes.

enrichment_database

Name of the database for enrichment. User can specify one or more database names from enrichR::listEnrichrDbs().

Value

enrichment_result a list of data.frames containing enrichment output and a list of plots of top 10 significant genesets.

See also