Preprocessing steps for Liger dimensionality reduction

Split merged object into multiple sce objects and extract sparse matrices:

liger_preprocess(
  sce,
  k,
  unique_id_var = "manifest",
  take_gene_union = F,
  remove.missing = T,
  make.sparse = T,
  num_genes = 3000,
  combine = "union",
  keep_unique = F,
  capitalize = F,
  do_plot = F,
  cex_use = 0.3,
  use_cols = T,
  ...
)

Arguments

sce
SingleCellExperiment object or merged objects

k
Inner dimension of factorization (number of factors).

Make a Liger object:

remove.missing
Whether to remove cells from scale.data with no gene expression (default TRUE)

Remove cells/genes with no expression across any genes/cells:

combine
How to combine variable genes across experiments. Either "union" or "intersect". (default "union")

capitalize
Capitalize gene names to match homologous genes (ie. across species) (default FALSE)

raw.data
List of expression matrices (gene by cell) each from an sce object.

take.gene.union
Whether to fill out raw.data matrices with union of genes across all datasets (filling in 0 for missing data) (requires make.sparse=T) (default FALSE).

num.genes
Number of genes to find for each dataset. Set to 3000 as default.

keep.unique
Keep genes that occur (i.e., there is a corresponding column in raw.data) only in one dataset (default FALSE)

do.plot
Display log plot of gene variance vs. gene expression for each dataset. Selected genes are plotted in green. (default FALSE)

cex.use
Point size for plot.

Scale genes by root-mean-square across cells:

use.cols
Treat each column as a cell (default TRUE)

Value

liger preprocessed object.