Functional enrichment analysis using enrichR

Performs impacted pathway analysis with a list of genes.

pathway_analysis_enrichr(
  gene_file = NULL,
  enrichment_database = c("GO_Molecular_Function_2018", "GO_Cellular_Component_2018",
    "GO_Biological_Process_2018", "MGI_Mammalian_Phenotype_2017", "ChEA_2016",
    "KEGG_2019_Human", "WikiPathways_2019_Human", "Reactome_2016",
    "Allen_Brain_Atlas_down", "Allen_Brain_Atlas_up"),
  is_output = FALSE,
  output_dir = "."
)

Arguments

gene_file
A data frame or the path of a .tsv file containing a list of genes, their fold-change, p-value and adjusted p-value. Column names should be gene, logFC, pval and padj respectively.

enrichment_database
Name of the database for enrichment. User can specify one or more database names from enrichR::listEnrichDbs().

is_output
If TRUE a folder will be created and results of enrichment analysis will be saved otherwise a R list will be returned. Default FALSE

output_dir
Path for the output directory. Default is current dir.

Value

enrichment_result a list of data.frames containing enrichment output and a list of plots of top 10 significant genesets.

See also

Examples

set.seed(1234)
enrichment_result <- pathway_analysis_enrichr(
  gene_file = paste(system.file("extdata", package = "scFlowData"), "/",
    "de_result_table.tsv",
    sep = ""
  ),
  enrichment_database = "GO_Molecular_Function_2018",
  is_output = FALSE
)

#> Error: File not found.